lil refactir
